Moved Initialization part from Application to Plugin in Android
