+++ android/src/main/java/com/rudderstack/sdk/flutter/RudderSdkFlutterPlugin.java
@@ -2,6 +2,7 @@ package com.rudderstack.sdk.flutter;
 
 import android.text.TextUtils;
 import androidx.annotation.NonNull;
+import android.content.Context;
 import com.rudderstack.android.sdk.core.RudderClient;
 import com.rudderstack.android.sdk.core.RudderConfig;
 import com.rudderstack.android.sdk.core.RudderMessageBuilder;
@@ -9,6 +10,7 @@ import com.rudderstack.android.sdk.core.RudderOption;
 import com.rudderstack.android.sdk.core.RudderProperty;
 import com.rudderstack.android.sdk.core.RudderTraits;
 import com.rudderstack.android.sdk.core.RudderTraitsBuilder;
+import com.rudderstack.android.sdk.core.RudderIntegration;
 import io.flutter.embedding.engine.plugins.FlutterPlugin;
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
@@ -29,6 +31,8 @@ public class RudderSdkFlutterPlugin
   /// when the Flutter Engine is detached from the Activity
   private MethodChannel channel;
   private static RudderClient rudderClient;
+  private Context context;
+  private static List<RudderIntegration.Factory> integrationList;
 
   @Override
   public void onAttachedToEngine(
@@ -40,12 +44,29 @@ public class RudderSdkFlutterPlugin
         "rudder_sdk_flutter"
       );
     channel.setMethodCallHandler(this);
+    context = flutterPluginBinding.getApplicationContext();
+  }
+
+
+  public RudderClient initializeSDK(MethodCall call) {
+    Map<String, Object> argumentsMap = (Map<String, Object>) call.arguments;
+    String writeKey = (String) argumentsMap.get("writeKey");
+
+    RudderConfig config = getRudderConfig(
+      (Map<String, Object>) argumentsMap.get("config")
+    );
+    RudderClient rudderClient = RudderClient.getInstance(
+      context,
+      writeKey,
+      config
+    );
+    return rudderClient;
   }
 
   @Override
   public void onMethodCall(@NonNull MethodCall call, @NonNull Result result) {
     if (call.method.equals("initializeSDK")) {
-      rudderClient = RudderSdkFlutterApplication.initializeSDK(call);
+      rudderClient = initializeSDK(call);
       return;
     } else if (call.method.equals("identify")) {
       HashMap<String, Object> argumentsMap = (HashMap<String, Object>) call.arguments;
@@ -185,6 +206,33 @@ public class RudderSdkFlutterPlugin
     channel.setMethodCallHandler(null);
   }
 
+  public RudderConfig getRudderConfig(Map<String, Object> configMap) {
+    RudderConfig.Builder builder = new RudderConfig.Builder();
+    builder
+      .withDataPlaneUrl((String) configMap.get("dataPlaneUrl"))
+      .withFlushQueueSize((Integer) configMap.get("flushQueueSize"))
+      .withDbThresholdCount((Integer) configMap.get("dbCountThreshold"))
+      .withConfigRefreshInterval(
+        (Integer) configMap.get("configRefreshInterval")
+      )
+      .withLogLevel((Integer) configMap.get("logLevel"))
+      .withSleepCount((Integer) configMap.get("sleepTimeOut"))
+      .withTrackLifecycleEvents((Boolean) configMap.get("trackLifecycleEvents"))
+      .withRecordScreenViews((Boolean) configMap.get("recordScreenViews"))
+      .withControlPlaneUrl((String) configMap.get("controlPlaneUrl"));
+    if (integrationList != null) {
+      builder.withFactories(integrationList);
+    }
+    return builder.build();
+  }
+
+  public static void addIntegration(RudderIntegration.Factory integration) {
+    if (integrationList == null) {
+      integrationList = new ArrayList<>();
+    }
+    integrationList.add(integration);
+  }
+ 
   public RudderTraits getRudderTraitsObject(Map<String, Object> traitsMap) {
     RudderTraitsBuilder builder = new RudderTraitsBuilder();
     if (traitsMap.containsKey("address")) {
